### 2018-04-06 #663 @whedon
[PRE REVIEW]: pdfsearch: Search Tools for PDF Files
**Submitting author:** @lebebr01 (<a href="http://orcid.org/0000-0002-1265-8761">Brandon LeBeau</a>)
**Repository:** <a href="https://github.com/lebebr01/pdfsearch" target ="_blank">https://github.com/lebebr01/pdfsearch</a>
**Version:** v0.2.0
