### 2018-04-06 #663 @arfon
@lebebr01 - please take a look at the `paper.md` file and fix the metadata so the paper will compile.
