### 2018-04-06 c8f5ccb
Update paper.md metadata.
