+++ inst/joss/paper.md
@@ -8,7 +8,10 @@ tags:
 authors:
  - name: Brandon LeBeau
    orcid: 0000-0002-1265-8761
-   affiliation: University of Iowa
+   affiliation: 1
+affiliations:
+ - name: University of Iowa
+   index: 1
 date: 30 March 2018
 bibliography: paper.bib
 ---
