### 2018-04-06 #663 @lebebr01
I just pushed the changes to the `paper.md` file that I believe will fix the compilation errors.

Looking at the list of reviewers, I wonder if @lmullen would be willing and able to review.
