### 2018-04-09 #668 @jeroen
@lebebr01 what is the reason pdfsearch 0.2.0 is not on CRAN yet?
