### 2018-04-09 #668 @lebebr01
@jeroen was primarily waiting to submit to CRAN until after the review. I'm open to submitting prior to review and then submitting another version after the review if there are changes made.
