### 2018-04-09 lebebr01/pdfsearch#10 @jeroen
Platform specific split string?
If I run the example from the readme I see different output:

Is this expected? I noticed [you split by `\r\n`](https://github.com/lebebr01/pdfsearch/blob/master/R/split_pdf.r#L3) which is windows-specific I think? Did you test the package on non-windows machines?
