### 2018-04-09 lebebr01/pdfsearch#10 @lebebr01
This is not expected and not an issue I was aware of. I do not have a non-windows machine currently to run some local tests like this on.

When I wrote this portion of the code I was not aware of the tokenizers package. It would likely be more robust to use their `tokenize_lines` function to perform this action.

I will also brainstorm a unit test that would catch this behavior as well. I'm open to ideas for a good way to test this.
