### 2018-04-09 lebebr01/pdfsearch#10 @jeroen
Yes you should add a unit test for this. You can automatically run checks on linux and osx using travis.
