### 2018-04-09 #668 @jeroen
@lebebr01 can you ping me here when https://github.com/lebebr01/pdfsearch/issues/10 has been fixed, so I can proceed with testing and reviewing the package (on macos ;))
