### 2018-04-09 e8d8e90
Convert \r\n to tokenize_lines
