+++ R/convert_tokens.r
@@ -38,7 +38,7 @@ convert_tokens <- function(x, path = FALSE, split_pdf = FALSE,
     line_nums <- cumsum(x_list[[2]])
     x_lines <- x_list[[1]]
   } else {
-    x_lines <- unlist(strsplit(x, split = '\r\n'))
+    x_lines <- unlist(tokenizers::tokenize_lines(x))
     x_lines <- gsub("^\\s+|\\s+$", '', x_lines)
   }
   
