+++ NAMESPACE
@@ -8,4 +8,5 @@ export(run_shiny)
 importFrom(pdftools,pdf_text)
 importFrom(pdftools,pdf_toc)
 importFrom(tibble,tibble)
+importFrom(tokenizers,tokenize_lines)
 importFrom(tokenizers,tokenize_words)
