### 2018-04-09 lebebr01/pdfsearch#10 @lebebr01
e8d8e90c63a9730cd0e99ea8defe7ea5c5bc705a should fix issue.

Added unit test here to test for literal "\n" characters in result text: [test here](https://github.com/lebebr01/pdfsearch/blob/e8d8e90c63a9730cd0e99ea8defe7ea5c5bc705a/tests/testthat/test_structure.r#L81-L90). Open to other ways to test this behavior.
