### 2018-04-09 e3ff001
Bump version following bug fix.
