### 2018-04-09 #668 @lebebr01
@jeroen Issue should be fixed. All functions now use `tokenizers::tokenize_lines` instead of the literal `\r\n` splitting, thanks for catching this.

Also added unit test here to test for literal "\n" characters in result text: [test here](https://github.com/lebebr01/pdfsearch/blob/e8d8e90c63a9730cd0e99ea8defe7ea5c5bc705a/tests/testthat/test_structure.r#L81-L90). Open to other suggestions to test this.
