### 2018-04-10 lebebr01/pdfsearch#10 @jeroen
OK thanks. BTW you could reduce your dependency weight by calling `stringi::stri_split_lines()` and `stringi::stri_split_boundaries(x, type = "word")` directly rather than via tokenizers.
