### 2018-05-08 lebebr01/pdfsearch#11 @lmullen
Corrections/typos in JOSS article
Corrections for [JOSS review](https://github.com/openjournals/joss-reviews/issues/668)

- [x] Paragraph 1, line 3, comma after `ubiquitous`
- [x] Paragraph 1, line 5, no comma after `library`, add comma after `documents`
- [x] Paragraph 2, line 3, remove parentheses in `files(s)`, because if the `s` were removed the sentence would be ungrammatical.
- [x] The phrase `than is currently done` is ambiguous. Please explain what research practice you are improving on.
- [x] The sentence `including example keywords searches and discusses` is not parallel and therefore ungrammatical.
- [x] Please expand the caption in figure 1 to explain what the sample search was. What keywords were you searching for, and over a single file or over a directory?
