### 2018-05-08 lebebr01/pdfsearch#12 @lmullen
Change depends to imports
Issue from [JOSS review](https://github.com/openjournals/joss-reviews/issues/668)

The package currently uses `Depends` in the `DESCRIPTION` for its dependencies. Best practice is to use `Imports`, and then to refer to the function by its namespace: e.g., `somepackage::some_func()`. See Hadley Wickham's [R Packages](http://r-pkgs.had.co.nz/description.html).
