### 2018-05-08 lebebr01/pdfsearch#13 @lmullen
Corrections to vignette
Issue for [JOSS review](https://github.com/openjournals/joss-reviews/issues/668).

- [x] Please give the vignette a title.
- [x] It would also be good to rename the vignette file to something less generic, like `intro-to-pdfsearch` or so on.

See this image for what the vignette currently shows up as:
