### 2018-05-08 lebebr01/pdfsearch#14 @lmullen
Simplify directory search interface?
Issue for [JOSS review](https://github.com/openjournals/joss-reviews/issues/668).

Why is it necessary for the user to specify `full_names = TRUE` when using the `keyword_directory()` function? When I passed in an absolute path to a directory, the function returned an error from `normalizePaths()` unless I used that argument. It would seem like the default option could be made easier for the user.
