### 2018-05-08 lebebr01/pdfsearch#15 @lmullen
Explain package significance and show research uses
Issue for [JOSS review](https://github.com/openjournals/joss-reviews/issues/668).

This package and paper is currently lacking a "a clear statement of need that illustrates the purpose of the software." The draft paper does explain that this package could be used to make keyword searching more reproducible, and I buy that claim. But I am not aware of a research application for keyword searching beyond looking things up, which would not need to be reproducible. What kinds of research work require this package? I'm willing to be persuaded, of course, but at the moment I don't understand why I would want to use this package instead of, e.g., just using my operating system's search feature.

Ideally the author would provided a worked example with a genuine research problem, either replacing or augmenting the demo example in the vignette.

The statement of need could perhaps be addressed by also meeting this JOSS guideline: "Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it." Could the author please explain what research applications he, and if possible, others, are using this package for?

This explanation of the research purpose definitely needs to go in the JOSS paper, but could also be adapted for the README.
