### 2018-05-08 #668 @lmullen
@arfon I have completed my review and I am willing to come back to it once @lebebr01 has had a chance to respond to the issues posted above.
