### 2018-05-15 7169452
Rename vignette and add vignette title entry
