+++ vignettes/intro_to_pdfsearch.Rmd
@@ -4,7 +4,7 @@ author: "Brandon LeBeau"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Vignette Title}
+  %\VignetteIndexEntry{Getting Started with pdfsearch}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
