### 2018-05-15 cd514a8
Draft edits to the JOSS paper
