+++ inst/joss/paper.bib
@@ -7,14 +7,14 @@
   url = {https://www.R-project.org/},
 }
 @Manual{pdfsearch,
-    title = {pdfsearch: Search Tools for PDF Files},
+    title = {{pdfsearch: Search Tools for PDF Files}},
     author = {Brandon LeBeau},
     note = {R package version 0.2.0},
     url = {https://github.com/lebebr01/pdfsearch},
     year = {2018}
 }
 @Manual{pdftools,
-    title = {pdftools: Text Extraction, Rendering and Converting of PDF Documents},
+    title = {{pdftools: Text Extraction, Rendering and Converting of PDF Documents}},
     author = {Jeroen Ooms},
     year = {2017},
     note = {R package version 1.5},
