### 2018-05-15 lebebr01/pdfsearch#14 @lebebr01
This is a good suggestion. I altered the default to `full_names = TRUE` instead of the old `full_names = FALSE`. This should simplify code and hopefully user experience.
