### 2018-05-15 lebebr01/pdfsearch#12 @lebebr01
Made this change. I was not aware of the `Imports` section, however have been using the `package::func()` syntax.
