### 2018-05-15 lebebr01/pdfsearch#13 @lebebr01
Added the title, "Getting Started with pdfsearch" to the \VignetteIndexEntry command.

I also made the suggested change to the vignette filename. The filename is now `intro_to_pdfsearch.Rmd`. I stuck with underscores as this is what I tend to use for all filenames in the package.
