### 2018-05-15 fa466eb
Adjust full_names documentation
