### 2018-05-15 c58d103
Remove full_names mention from README
