### 2018-05-15 e5a0d98
further clarification of example package usage
