+++ inst/joss/paper.bib
@@ -20,3 +20,10 @@
     note = {R package version 1.5},
     url = {https://CRAN.R-project.org/package=pdftools},
 }
+@book{cooper2016,
+  title={Research synthesis and meta-analysis: A step-by-step approach},
+  author={Cooper, Harris},
+  volume={2},
+  year={2016},
+  publisher={Sage publications}
+}
