+++ inst/joss/paper.md
@@ -20,7 +20,7 @@ bibliography: paper.bib
 
 PDF files are common formats for reports, journal articles, briefs, and many other documents. PDFs are lightweight, portable, and easily viewed across operating systems. Even though PDF files are ubiquitous, extracting and finding text within a PDF can be time consuming and not easily reproducible. The pdftools R package [@pdftools], which uses the poppler C++ library to extract text from PDF documents, aids in the ability to import text data from PDF files to manipulate in R. The pdfsearch package [@pdfsearch] is an R package [@r-base] that extends the text extraction of pdftools to allow for keyword searching within a single PDF or a directory of PDF files.
 
-The pdfsearch package can aid users in manipulation of text data from PDF files in R and may also improve the reproducibility of the extraction and manipulation tasks. Users can search for keywords within PDF files where the location of the match and the raw text from match are returned. This aspect of searching for keywords may be most useful for those conducting research syntheses or meta-analyses that are more reproducible and less time consuming than current practice. Current research synthesis or meta-analysis practice involves the reading of each article to search for the presence of certain terms, phrases, or statistical effect size information. The improved workflow with the pdfsearch package would allow those conducting research syntheses keyword searches to explore only the relavent areas of the PDF text surrounding the matching text. In addition, regular expressions could be written to search for statistical effect size information to extract automatically.
+The pdfsearch package can aid users in manipulation of text data from PDF files in R and may also improve the reproducibility of the extraction and manipulation tasks. Users can search for keywords within PDF files where the location of the match and the raw text from match are returned. This aspect of searching for keywords may be most useful for those conducting research syntheses or meta-analyses [@cooper2006] that are more reproducible and less time consuming than current practice. Current research synthesis or meta-analysis practice involves the reading of each article to search for the presence of certain terms, phrases, or statistical effect size information. The improved workflow with the pdfsearch package would allow those conducting research syntheses keyword searches to explore only the relavent areas of the PDF text surrounding the matching text. In addition, regular expressions could be written to search for statistical effect size information to extract automatically. As an example, the package is currently being used to explore the evolution of statistical software and quantitative methods used in published research. This process involves getting PDF files for published research articles and using pdfsearch to search for specific software and quantitative methods keywords within the research articles. Prevalence of the terminology is explored using research synthesis methods [@cooper2006].
 
 The package vignette includes more information on this package. Included in the vignette are keyword searches within PDF documents and an exploration of the output from the package. The vignette also discusses limitations of the package. Below is example output of one of these examples. Link to the package release can be found here: <https://doi.org/10.5281/zenodo.1210279>
 
