### 2018-05-15 b2d74ed
adjust to tibble::is.tibble in test
