### 2018-05-17 b501702
closes #11 #15
