+++ inst/joss/paper.bib
@@ -20,10 +20,30 @@
     note = {R package version 1.5},
     url = {https://CRAN.R-project.org/package=pdftools},
 }
-@book{cooper2016,
+@book{cooper2017,
   title={Research synthesis and meta-analysis: A step-by-step approach},
   author={Cooper, Harris},
   volume={2},
-  year={2016},
+  year={2017},
   publisher={Sage publications}
+}
+@misc{lebeau2018,
+ author = {Brandon LeBeau and Ariel M Aloe},
+ title = {Evolution of Statistical Software and Quantitative Methods},
+ note = {Paper accepted to the Joint Statistical Meetings, Vancouver, BC (2018)},
+ year = {2018},
+ month = {July}
+}
+@article{2015arXiv,
+   author = {{Guo}, Y. and {Deng}, A.},
+    title = "{Flexible Online Repeated Measures Experiment}",
+  journal = {ArXiv e-prints},
+archivePrefix = "arXiv",
+   eprint = {1501.00450},
+ primaryClass = "stat.AP",
+ keywords = {Statistics - Applications, G.3},
+     year = 2015,
+    month = jan,
+   adsurl = {http://adsabs.harvard.edu/abs/2015arXiv150100450G},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
