### 2018-05-17 lebebr01/pdfsearch#15 @lebebr01
Edited the JOSS paper to include expanded discussion of how I feel this paper could be useful for research synthesis or meta-analysis projects. Also included description of current research using the package to explore evolution of statistical software and quantitative methods in published social science research that is going to be presented at JSM this summer.
