### 2018-05-17 lebebr01/pdfsearch#11 @lebebr01
Thank you for these comments/edits. These have been adjusted in the JOSS paper.
