### 2018-05-17 009131d
Add research section to README #15
