### 2018-05-17 #668 @lebebr01
First, a belated thank you to @jeroen for your review.

Secondly, thank you to @lmullen for your review. I have made the suggested edits related to the code and vignette issues identified above.

I also have expanded on the significance and usage of the package in research settings in the JOSS paper and README. My thoughts are that this package would be of primary interest to those doing research syntheses or meta-analyses. I also present description of an ongoing research project using the package to explore the evolution of statistical software and quantitative methods used in published social science research that will be presented at JSM this summer. I'd be interested in your thoughts on whether this expanded discussion is compelling.

Changes and a few additional comments can be seen in the issues above. I also tried to link major changes via commits in the issues above.
