### 2018-05-24 fa2f044
Add shiny demo of interactive selection
