### 2018-05-29 #668 @lmullen
Two minor comments on the revisions to the JOSS paper.

"instead of searching the entire text of a document specific portions of the text": comma after document.

"a preprint will be posted here." I'm not sure from the context where "here" is.

Other than that I am fine with the revisions, which I have reviewed.
