### 2018-05-30 78607d9
Update README grammar for JOSS review
