### 2018-06-19 #668 @lmullen
Good by me.

On Tue, Jun 19, 2018 at 7:14 PM Karthik Ram <notifications@github.com>
wrote:

> @lebebr01 <https://github.com/lebebr01> Just returning from vacation. I'm
> happy to accept (assuming @lmullen <https://github.com/lmullen> is not
> waiting on anything else for the final checks). Please archive your
> software on Zenodo and post the DOI here so we can proceed with next steps.
> 🙏
>
