### 2018-06-20 6492f58
remove zenodo link joss paper
