### 2018-06-20 #668 @lebebr01
@karthik I've archived the latest version to Zenodo and the DOI is  10.5281/zenodo.1294067 (https://doi.org/10.5281/zenodo.1294067).

Thanks again for the reviews @jeroen and @lmullen.
