### 2018-07-07 #668 @arfon
@karthik are we good to accept here?
