### 2018-07-08 #668 @karthik
Yes good to accept. Was waiting on PDF and forgot to ping you.
